fix texture issue for the 3D sample
The issue was that the texture parameter GL_TEXTURE_WRAP_R is set to GL_REPEAT by default.
The texture R coordinate was set to 1.0, so it blended with the neighbour, which was the
R coordinate 0.0, because the texture wrapped around. Seting the texture wrap to
CLAMP_TO_EDGE means that it doesn't do that.